2023-02-13 [yj] /exhList 인터셉터 추가

+++ GDJ56_Exhibition_final/src/main/webapp/WEB-INF/views/news/newsView.jsp
@@ -45,7 +45,7 @@ div.newscontainer{
                 	onclick="fn_filedownload('${file.originalFilename}','${file.renamedFilename}');">${file.originalFilename }</button>
             </c:forEach><br><br>
             <a href="${path }/newsUpdateWrite.do?newsNo=${param.newsNo}" class="btn btn-sm btn-dark btn-block" role="button">수정</a>
-            <a href="${path }/newsDelete.do?newsNo=${param.newsNo}" class="btn btn-sm btn-dark btn-block " role="button">삭제</a>
+            <a href="${path }/newsDelete.do?newsNo=${param.newsNo}" class="btn btn-sm btn-dark btn-block" onclick="if(!confirm('삭제 하시겠습니까?'))" role="button">삭제</a>
         </div>
         <div class="card-body">
             <a href='${path }/newslist.do' class="btn btn-sm btn-dark btn-block" role="button">목록으로</a>
